Devel spartan782 (#25)
* initial

started building ansible Playbook to create iso.

* Successful run with 2 external module imports

Actually runs 2/3rds of offline-snapshot.sh utilizing an ansible playbook.

Currently downloads the entry dependancy tree and downloads any missing packages. Then proceeds to sign them with supplied key and password.

* Add url downloads and File signing

* Fix repobuilder library

Moved group packages below user check to make sure that if not root that we were working in a temp cache.

* Add group_vars

Added group vars and changed common role to use the vars

* Update readme

Updated with steps required when using the ansible role to create the offline snapshot

* Update bootstrap.sh

Add ansible as dependency in bootstrap

* INIT Concourse script

Script used for concourse pipeline

* Update master ISO

Updated to use the ansible portion of offline snapshot

* fix concourse script

* attempted concourse fix

Round 2 fix?

* Round 3 Concourse

* made concourse executable

* concourse working dir

* Update concourse script

add dvd input and output

* Change branches to devel

Change branches for bro scripts and dashboards to devel

* Make concourse folder

move concourse stuff to its own home.

* Make executable

* Concourse error not found

Attempt at fixing concourse error of not being able to locate the ansible script

* Concourse ansible error

Another iteration of conourse not being able to locate ansible scripts

* Concourse testing

Still troubleshooting concourse

* change execution of create iso

* Testing Concourse

Move up 2 directories

* Concourse Possibly fixed

Tests seemed successful. This should work?

* Conourse add parent dir flag

* Fix jinja2

fixed jinaja 2 in ansible playbook that needed quotes

* change ansible host to localhost

* Fix directory issue

script fails to execute if you called the script from a different location than where it is sitting currently

* Add skip var for mcopy

* Fix Rock Cache dir

* Add additional override to ansible

* Add local pulp server to yum conf

* Put local yums in the right file

* Attempt fix for rock cache dir

* change efi setting in master iso script

* Minor variable tweak

* Revert "Minor variable tweak"

This reverts commit 42ef8f949005278a6a5753b5337419a97c45dbea.

* Add gpg signing option to master iso

Reworked this shell script to allow optional flags should you want to pass a gpg key name, password that is already imported or even an option to import a gpg key to be used in creationg of iso.

* Add concourse pipeline

* Update concourse script to support gpg signing

Still a work in progress as this is waiting on the creation of a vault server to provide secrets.

* Tweak vars

* Smoke test Concourse GPG sign

First test round of having concourse sign things and use vault to pull the secrets

* Fix script errors

* Attempting Concourse Variables in script

Concourse only allows secrets to be in certain area's the closest I can get them to the script is in the ENV. Testing to see if I can pass environment vars to a script.

* Concourse execute with Vars

Changed to accomidate the passing of variables to concourse.

* Allow spaces in variables

* Fix syntax errror

* Update readme

Added in usage for new master-iso script as well as how to utilize the ansible portion of creating offline snapshots.

* Additional space syntax error fixes

* Prevent password from being logged

* Hide password output

* Fix set -x options

* Change unset to set +x

* Fix Jinja2 variable

* Update concourse pipeline

Current concourse pipeline for building iso's and using vault secrets

* Fix syntax error

* Remove duplicate method call

* change ansible when statement

* Fix GPG key import

By also importing the GPG key into RPM this should allow for the packages to be signed

* Concourse verbose ansible logging

* Enable GPG macro's

Add the gpg variable to the macro's section for signing.

Add --batch and --passphrase options to gpg2

* Install rpm-sign

* Add Bash safety quotes

With out double quotes around the whole argument and single quotes around the variable bash will eat the quotes which will ultimately cut the variable if it has spaces.

* Repodata signing from upstream

* Add output directory

* init Concourse Upload

Script should execute pulp admin commands to upload iso files created from the pipeline.

* Update git ignore

* Fix Output directory of ISO

* Update pulp admin conf

* make upload script executable

* Update Concourse Pipeline

* Add pulp cert to upload

* Create Nightly, Release, Manual pipelines

Created a pipeline for different concourse pipelines. One for Manual builds, one for Nightly Builds and One for releases.

* Fix Passphrase

* Fix Upload Hostname

* Change home Var

* Update mirror certificate